pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Production-grade conversational RAG backend"
requires-python = ">=3.11,<3.12"

dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.1",
    "sqlalchemy==2.0.27",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "pydantic==2.6.1",
    "qdrant-client==1.7.3",
    "python-dotenv==1.0.1",
    "httpx==0.27.0",
    "loguru==0.7.2",
    "greenlet==3.0.3",
    "pydantic-settings==2.2",
    "openai>=1.0.0,<2.0.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["migrations*", "tests*"]
